requirements.txt
flask
gunicorn
requests
beautifulsoup4
python-dotenv
langchain
langchain-core
langchain-groq
langchain-huggingface
langchain-pinecone
pinecone-client
langchain-community